Feat: Integrate API communication with frontend, add methods to fetch all post and get post by ID
