+++ src/utils/global-data.js
