+++ tailwind.config.js
@@ -1,9 +1,9 @@
 module.exports = {
   mode: 'jit',
   content: [
-    './pages/**/*.{js,ts,jsx,tsx}',
-    './components/**/*.{js,ts,jsx,tsx}',
+    './src/pages/**/*.{js,ts,jsx,tsx}',
+    './src/components/**/*.{js,ts,jsx,tsx}',
   ],
   darkMode: 'class', // or 'media' or 'class'
-  presets: [require('./utils/tailwind-preset')],
+  presets: [require('./src/utils/tailwind-preset')],
 };
