--- utils/mdx-utils.js
@@ -1,19 +1,0 @@
-import { api } from '../services/api'
-
-export const getPosts = async () => {
-    const {data} = await api.get('/posts'); 
-
-    if(data){
-        return data;
-    }
-
-    return []
-}
-
-export const getPostBySlug = async (id) => {
-
-    //TODO: BUSCAR UM POST EM ESPECIFICO.
-    //const {data} = await api.get(`/post?id=eq.${id}`)
-
-    return {}
-}
